# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateOrderItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  models.CreateTelegramOrderRequest:
    properties:
      customer_id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CreateOrderItemRequest'
        type: array
      merchant_id:
        type: string
    required:
    - customer_id
    - items
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
    type: object
  models.Merchant:
    properties:
      merchant_id:
        type: string
      merchant_name:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      customer_id:
        type: string
      id:
        type: string
      order_date:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        $ref: '#/definitions/models.OrderStatus'
      total_price:
        type: number
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
    type: object
...
  models.OrderStatus:
    enum:
    - pending
    - confirmed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusConfirmed
    - OrderStatusCancelled
  models.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      public_id:
        type: string
      stock:
        type: integer
      user_id:
        type: string
    type: object
...
  utils.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
  utils.ResponsePaginate:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing users, products, and orders
  termsOfService: http://swagger.io/terms/
  title: Imphnen API
  version: "0.1"
paths:
...
  /telegram/customers/{customer_id}/orders:
    get:
      consumes:
      - application/json
      description: Get all orders for a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponsePaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Order'
                  type: array
                meta:
                  $ref: '#/definitions/utils.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List customer orders (Telegram bot)
      tags:
      - Telegram
...
  /telegram/merchants:
    get:
      consumes:
      - application/json
      description: Get all merchants/users with their ID and store name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Merchant'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all merchants (for Telegram bot)
      tags:
      - Telegram
  /telegram/merchants/{merchant_id}/products:
    get:
      consumes:
      - application/json
      description: Get all products available from a specific merchant
      parameters:
      - description: Merchant User ID
        in: path
        name: merchant_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponsePaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
                meta:
                  $ref: '#/definitions/utils.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List products by merchant (for Telegram bot)
      tags:
      - Telegram
  /telegram/orders:
    post:
      consumes:
      - application/json
      description: Create a new order for a customer via telegram bot
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.CreateTelegramOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create order for customer (Telegram bot)
      tags:
      - Telegram
  /telegram/orders/{order_id}:
    delete:
      consumes:
      - application/json
      description: Delete a pending or cancelled order
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete customer order (Telegram bot)
      tags:
      - Telegram
  /telegram/orders/{order_id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel an order and restore stock
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Cancel customer order (Telegram bot)
      tags:
      - Telegram
  /telegram/orders/{order_id}/confirm:
    patch:
      consumes:
      - application/json
      description: Accpet an order and change stock
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Accept customer order (Telegram bot)
      tags:
      - Telegram
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...